Remove the Garbage struct (closes #12)
